update scaffolds implementation to count ring systems instead of rings
