+++ src/generate_scaffolds.py
@@ -11,152 +11,11 @@ badapple database.
 
 import argparse
 
-import scaffoldgraph as sg
-from loguru import logger
 from rdkit import Chem
-from scaffoldgraph.core.fragment import (
-    get_annotated_murcko_scaffold,
-    get_murcko_scaffold,
-)
-from scaffoldgraph.core.scaffold import Scaffold
-from scaffoldgraph.io import *
 
 from utils.custom_logging import get_and_set_logger
 from utils.file_utils import close_file, get_csv_writer
-
-
-def canon_smiles(mol: Chem.Mol, kekule=False):
-    # beware of oscillating SMILES when using kekule=True !
-    # example:
-    # O=C1C=CN2CCCC3=CC=CC1=C32 => O=C1C=CN2CCCC3=C2C1=CC=C3 => O=C1C=CN2CCCC3=CC=CC1=C32
-    try:
-        canon_smiles = Chem.MolToSmiles(mol, canonical=True, kekuleSmiles=kekule)
-        # converting again... why:
-        # in edge cases there is some information loss when converting to SMILES
-        # by converting twice, we ensure consistency with the PostgreSQL DB and RDKit
-        # example of what happens without converting twice:
-        # >>> scaf1 = Chem.MolFromSmiles("C1=NC(c2ccccc2)=NP=N1")
-        # >>> scaf2 = Chem.MolFromSmiles("c1ccc(-c2ncnpn2)cc1")
-        # >>> Chem.MolToSmiles(scaf1) == Chem.MolToSmiles(scaf2)
-        # True
-        # basically in ScaffoldGraph scaf1 and scaf2 were considered to be different and output different canonSMILES,
-        # even though when we reconstruct the scaffolds with these SMILES they are the same
-        # this issue happens even with kekule=True, for example:
-        # >>> scaf1 = Chem.MolFromSmiles("C(=NC1=C(N2CCCCC2)C=CC=C1)C1=CNN=C1")
-        # >>> scaf2 = Chem.MolFromSmiles("C(=NC1=CC=CC=C1N1CCCCC1)C1=CNN=C1")
-        # >>> Chem.MolToInchi(scaf1) == Chem.MolToInchi(scaf2)
-        # [15:23:11] WARNING: Omitted undefined stereo
-        # [15:23:11] WARNING: Omitted undefined stereo
-        # True
-        # >>> Chem.MolToSmiles(scaf1,kekuleSmiles=True) == Chem.MolToSmiles(scaf2,kekuleSmiles=True)
-        # True
-        canon_smiles = Chem.MolToSmiles(
-            Chem.MolFromSmiles(canon_smiles), canonical=True, kekuleSmiles=kekule
-        )
-        return canon_smiles
-    except:
-        # this is an extreme edge case, but certain SMILES output by ScaffoldGraph do not play nice with other functions in RDKit
-        # (e.g., Chem.MolFromSmiles)
-        # only example seen so far: [c-]1cccc1.c1cc(N2CCCC2)[c-]2[cH-][cH-][cH-][c-]2n1
-        # (this was from CID 16717706)
-        original_smiles = Chem.MolToSmiles(mol)
-        return original_smiles
-
-
-class CustomHierS(sg.HierS):
-    """
-    This is a slightly modified version of the original HierS algorithm from ScaffoldGraph. it uses the following changes:
-    1) Includes molecules with no top-level scaffold in the graph.
-    2) Supports multiple identifier types, rather than only canonical aromatic SMILES
-    """
-
-    def __init__(self, *args, identifier_type="canon_smiles", **kwargs):
-        super().__init__(*args, **kwargs)
-        # Track scaffolds that couldn't be Kekulized
-        # (these structures are invalid for RDKit PostgreSQL cartridge)
-        self.non_kekule_scaffolds = set()
-        self.identifier_type = identifier_type
-        if identifier_type == "canon_smiles":
-            self.hash_func = canon_smiles
-        elif identifier_type == "kekule_smiles":
-            self.hash_func = lambda mol: canon_smiles(mol, kekule=True)
-        elif identifier_type == "inchi":
-            self.hash_func = Chem.MolToInchi
-        else:
-            raise ValueError(f"Unrecognized identifier_type: {identifier_type}")
-
-    def _process_no_top_level(self, molecule):
-        """Private: Process molecules with no top-level scaffold.
-        Modified from original code so that molecules with no top-level
-        scaffold are still included in the graph.
-        Parameters
-        ----------
-        molecule : rdkit.Chem.rdchem.Mol
-            An rdkit molecule determined to have no top-level scaffold.
-        """
-        name = molecule.GetProp("_Name")
-        logger.info(f"No top level scaffold for molecule: {name}")
-        self.graph["num_linear"] = self.graph.get("num_linear", 0) + 1
-        self.add_molecule_node(
-            molecule,
-        )
-        return None
-
-    def _initialize_scaffold(self, molecule, init_args):
-        """Initialize the top-level scaffold for a molecule.
-        Modified from the original code to Kekulize the scaffold.
-
-        Initialization generates a murcko scaffold, performs
-        any preprocessing required and then adds the scaffold
-        node to the graph connecting it to its child molecule.
-        This process can be customised in subclasses to modify
-        how a scaffold is initialized.
-
-        Parameters
-        ----------
-        molecule : rdkit.Chem.rdchem.Mol
-            A molecule from whicg to initialize a scaffold.
-        init_args : dict
-            A dictionary containing arguments for scaffold
-            initialization and preprocessing.
-
-        Returns
-        -------
-        scaffold : Scaffold
-            A Scaffold object containing the initialized
-            scaffold to be processed further (hierarchy
-            generation).
-
-        """
-        scaffold_rdmol = get_murcko_scaffold(molecule)
-        if scaffold_rdmol.GetNumAtoms() <= 0:
-            return self._process_no_top_level(molecule)
-        scaffold_rdmol = self._preprocess_scaffold(scaffold_rdmol, init_args)
-        scaffold = Scaffold(scaffold_rdmol)
-        # CHANGE: override default hash_func
-        scaffold.hash_func = self.hash_func
-        # END CHANGE
-        annotation = None
-        if init_args.get("annotate") is True:
-            annotation = get_annotated_murcko_scaffold(molecule, scaffold_rdmol, False)
-        self.add_scaffold_node(scaffold)
-        self.add_molecule_node(molecule)
-        self.add_molecule_edge(molecule, scaffold, annotation=annotation)
-        return scaffold
-
-    def _hierarchy_constructor(self, child):
-        parents = (p for p in self.fragmenter.fragment(child) if p)
-        for parent in parents:
-            # CHANGE: use consistent hash_func as in _initialize_scaffold
-            parent.hash_func = self.hash_func
-            # END CHANGE
-            if parent in self.nodes:
-                self.add_scaffold_edge(parent, child)
-            else:
-                self.add_scaffold_node(parent)
-                self.add_scaffold_edge(parent, child)
-                if parent.ring_systems.count > 1:
-                    self._hierarchy_constructor(parent)
+from utils.hiers import CustomHierS
 
 
 def parse_args(parser: argparse.ArgumentParser):
@@ -203,8 +62,8 @@ def parse_args(parser: argparse.ArgumentParser):
         "--max_rings",
         type=int,
         required=False,
-        default=20,
-        help="Maximum number of rings allowed in input compounds. Compounds with > max_rings will not be processed",
+        default=10,
+        help="Maximum number of ring systems allowed in input compounds. Compounds with > max_rings will not be processed",
     )
     parser.add_argument(
         "--identifier_type",
@@ -362,13 +221,15 @@ def write_outs(
 
 
 def main(args):
+    args_dict = vars(args)
+    logger.info(f"Running generate_scaffolds.py with the following args: {args_dict}")
     network = CustomHierS.from_smiles_file(
         file_name=args.i,
         header=args.iheader,
         delimiter=args.idelim,
         smiles_column=args.smiles_column,
         name_column=args.name_column,
-        ring_cutoff=args.max_rings,
+        ring_cutoff=args.max_rings,  # note that this is counting ring systems, not rings
         progress=True,
     )
     write_outs(
@@ -378,6 +239,14 @@ def main(args):
         args.o_scaf,
         args.o_mol2scaf,
         "\t",
+    )
+    logger.info(f"Total number of molecules in graph: {network.num_molecule_nodes}")
+    logger.info(f"Total number of scaffolds in graph: {network.num_scaffold_nodes}")
+    logger.info(
+        f"Total number of input molecules filtered (> max_rings): {network.graph["num_filtered"]}"
+    )
+    logger.info(
+        f"Total number of linear molecules (molecules with no scaffolds) in input: {network.graph["num_linear"]}"
     )
